+++ webextension/js/variousFeatures/contentScripts.js
@@ -5,6 +5,7 @@ import {getUserscriptData, setUserscriptData, writeClipboard} from "../classes/c
 import {errorToString} from "../utils/errorToString.js";
 import dateUtils from '../utils/dateUtils.js';
 import {getPanelPorts} from "../panelPort.js";
+import {slugify} from "../utils/slugify.js";
 
 // noinspection JSUnusedGlobalSymbols,JSUnusedLocalSymbols
 const znmDataApi = {
@@ -551,7 +552,7 @@ function onUserScriptMessage(message, sender, sendResponse) {
 	}
 }
 
-function userScriptApiLoader(context, dateUtils) {
+function userScriptApiLoader(context, dateUtils, slugify) {
 	chrome.runtime.sendMessage({ type: 'user_script_executed', userScriptsId: context.fileName }).catch(console.error);
 	const call = async function userScriptApiCall() {
 		const [callName, ...args] = arguments;
@@ -634,6 +635,7 @@ function userScriptApiLoader(context, dateUtils) {
 	const znmApi = new Proxy({
 		...context,
 		date: dateUtils,
+		slugify: slugify,
 		on(eventName, listener) {
 			if (!(eventName in listeners)) {
 				listeners[eventName] = [];
@@ -1029,7 +1031,7 @@ class ContentScripts {
 			id: userScript.fileName,
 			runAt: userScript.runAt? userScript.runAt.replace(/-/g, '_') : undefined,
 			js: [
-				{ code: `'use strict';const znmApi_${uniqSuffix} = ${userScriptApiLoader.toString()}(${JSON.stringify(context)}, ${dateUtilsString});\n(function(znmApi, unsafeWindow, window${additionalParams}){ ${userScript.script} }).call(znmApi_${uniqSuffix}, znmApi_${uniqSuffix}, window, undefined${additionalValues});` },
+				{ code: `'use strict';const znmApi_${uniqSuffix} = ${userScriptApiLoader.toString()}(${JSON.stringify(context)}, ${dateUtilsString}, ${slugify.toString()});\n(function(znmApi, unsafeWindow, window${additionalParams}){ ${userScript.script} }).call(znmApi_${uniqSuffix}, znmApi_${uniqSuffix}, window, undefined${additionalValues});` },
 			],
 			matches: userScript.match ?? [],
 			excludeMatches: userScript.excludeMatches ?? [],